stock_ws/requirements.txt
lxml
suds-jurko

